Feat: Add article routing based on static MD files
The articles section of the website will run off of static Markdown
files in a data directory.

This commit adds Next.js routing for the /articles and /article/[slug]
paths.

At build time, the content service reads all files in the
WORKSPACE_ROOT/data/articles directory using Effect.

This commit does not parse the article files and display the content.
This will be done in a future commit.